Restructure/refactor to prepare for a11y for a module
