+++ packages/a11y/src/shared/clear.js
